Adding the required functionality to gmode_lineBE's GLineBETranslator
Due to different requirements for the GLineBE translator to use 'AI_wave', 'BE_wave_AO_0', 'BE_wave_AO_1', 'BE_wave_train', 'BE_wave' etc. for the given dataset. We are creating another filter that can be used as per our requirements.
Some more changes might be pushed into it based on the type of data and vectors we want in our HDF5 file later on.